refactor(daemon): give RPC follower lifecycle its own owner
The complexity ratchet is owner-locked to only move down, and the follower
guard pushed server.ts past its baseline. The connection lifecycle, the
disconnected-follower cleanup, and the guarded send now live in a small
dedicated RpcFollowers owner, which puts server.ts below its baseline again
without changing behavior. The regression test reaches the follower set
through the new owner.
